separated prover and verifier
