+++ stwo_cairo_prover/crates/prover/src/cairo_air/verifier.rs
@@ -1,0 +1,67 @@
+use num_traits::Zero;
+use stwo_cairo_common::memory::LOG_MEMORY_ADDRESS_BOUND;
+use stwo_prover::core::channel::MerkleChannel;
+use stwo_prover::core::fields::qm31::SecureField;
+use stwo_prover::core::pcs::{CommitmentSchemeVerifier, PcsConfig};
+use stwo_prover::core::prover::{verify, VerificationError};
+use thiserror::Error;
+
+use super::CairoProof;
+use crate::cairo_air::air::{lookup_sum, CairoComponents, CairoInteractionElements};
+use crate::components::memory_address_to_id::component::MEMORY_ADDRESS_TO_ID_SPLIT;
+
+pub fn verify_cairo<MC: MerkleChannel>(
+    CairoProof {
+        claim,
+        interaction_claim,
+        stark_proof,
+    }: CairoProof<MC::H>,
+    pcs_config: PcsConfig,
+) -> Result<(), CairoVerificationError> {
+    // Auxiliary verifications.
+    // Assert that ADDRESS->ID component does not overflow.
+    assert!(
+        (1 << claim.memory_address_to_id.log_size) * MEMORY_ADDRESS_TO_ID_SPLIT
+            <= (1 << LOG_MEMORY_ADDRESS_BOUND)
+    );
+
+    let channel = &mut MC::C::default();
+    let commitment_scheme_verifier = &mut CommitmentSchemeVerifier::<MC>::new(pcs_config);
+
+    let log_sizes = claim.log_sizes();
+
+    // Preproccessed trace.
+    commitment_scheme_verifier.commit(stark_proof.commitments[0], &log_sizes[0], channel);
+
+    claim.mix_into(channel);
+    commitment_scheme_verifier.commit(stark_proof.commitments[1], &log_sizes[1], channel);
+    let interaction_elements = CairoInteractionElements::draw(channel);
+
+    // Verify lookup argument.
+    if lookup_sum(&claim, &interaction_elements, &interaction_claim) != SecureField::zero() {
+        return Err(CairoVerificationError::InvalidLogupSum);
+    }
+    interaction_claim.mix_into(channel);
+    commitment_scheme_verifier.commit(stark_proof.commitments[2], &log_sizes[2], channel);
+
+    let component_generator =
+        CairoComponents::new(&claim, &interaction_elements, &interaction_claim);
+    let components = component_generator.components();
+
+    // Verify stark.
+    verify(
+        &components,
+        channel,
+        commitment_scheme_verifier,
+        stark_proof,
+    )
+    .map_err(CairoVerificationError::Stark)
+}
+
+#[derive(Error, Debug)]
+pub enum CairoVerificationError {
+    #[error("Invalid logup sum")]
+    InvalidLogupSum,
+    #[error("Stark verification error: {0}")]
+    Stark(#[from] VerificationError),
+}
