+++ stwo_cairo_prover/crates/recursion_bridge/src/proof_generation/mod.rs
@@ -4,7 +4,8 @@ mod tests {
 
     use cairo_lang_casm::casm;
     use stwo_cairo_adapter::plain::input_from_plain_casm_with_step_limit;
-    use stwo_cairo_prover::cairo_air::{prove_cairo, verify_cairo, ProverConfig};
+    use stwo_cairo_prover::cairo_air::prover::{prove_cairo, ProverConfig};
+    use stwo_cairo_prover::cairo_air::verifier::verify_cairo;
     use stwo_prover::core::pcs::PcsConfig;
     use stwo_prover::core::vcs::blake2_merkle::Blake2sMerkleChannel;
