+++ stwo_cairo_prover/crates/run_and_prove/src/main.rs
@@ -4,10 +4,10 @@ use std::process::ExitCode;
 use clap::Parser;
 use stwo_cairo_adapter::plain::adapt_finished_runner;
 use stwo_cairo_adapter::vm_import::VmImportError;
-use stwo_cairo_prover::cairo_air::{
-    default_prod_prover_parameters, prove_cairo, verify_cairo, CairoVerificationError,
-    ProverConfig, ProverParameters,
+use stwo_cairo_prover::cairo_air::prover::{
+    default_prod_prover_parameters, prove_cairo, ProverConfig, ProverParameters,
 };
+use stwo_cairo_prover::cairo_air::verifier::{verify_cairo, CairoVerificationError};
 use stwo_cairo_utils::binary_utils::run_binary;
 use stwo_cairo_utils::vm_utils::{run_vm, VmArgs, VmError};
 use stwo_prover::core::prover::ProvingError;
